Make `compass setup` create a custom test suite
This is convenient for running several test cases in one command
even if they aren't part of a predefined test suite.  Test cases
are run in the order they are supplied to `compass setup`.

Some functionality has been moved from the ``suite`` module into
``setup`` and ``clean`` to facilitate setting up and cleaning up
custom test suites in the same way as predefined ones.

A small bug has been fixed in determining the number of cores
that a suite requires.  Previously, all steps were used.  With
the fix, only ``steps_to_run`` are used.
